faceAcecom-v2/requirements.txt
matplotlib==3.1.1
tensorflow
numpy==1.18.1
opencv_python==4.2.0.34
Pillow==7.1.2
streamlit
streamlit-webrtc 
opencv-python-headless
pydub
python-dotenv 